pyproject.toml
[project]
name = "maptoposter"
version = "0.3.0"
description = "Generate beautiful, minimalist map posters for any city in the world."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Ankur Gupta", email = "originalankur@github.com"}
]
keywords = ["map", "poster", "osm", "openstreetmap", "art", "visualization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Multimedia :: Graphics",
    "Typing :: Typed",
]

dependencies = [
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "fonttools==4.61.1",
    "geographiclib==2.1",
    "geopandas==1.1.2",
    "geopy==2.4.1",
    "idna==3.11",
    "kiwisolver==1.4.9",
    "lat-lon-parser==1.3.1",
    "matplotlib==3.10.8",
    "networkx==3.6.1",
    "numpy==2.3.3",
    "osmnx==2.0.7",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==12.1.0",
    "pyogrio==0.12.1",
    "pyparsing==3.3.1",
    "pyproj==3.7.2",
    "pyyaml==6.0.2",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "requests==2.32.5",
    "scipy==1.16.3",
    "shapely==2.1.2",
    "six==1.17.0",
    "tqdm==4.67.1",
    "tzdata==2025.3",
    "urllib3==2.6.3",
]

[project.scripts]
maptoposter-cli = "maptoposter.cli:main"

[project.urls]
Homepage = "https://github.com/originalankur/maptoposter"
Source = "https://github.com/originalankur/maptoposter"
"Bug Tracker" = "https://github.com/originalankur/maptoposter/issues"
Changelog = "https://github.com/originalankur/maptoposter/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
maptoposter = ["themes/*.json", "fonts/*.ttf", "py.typed"]

[tool.pytest.ini_options]
addopts = "--cov=maptoposter --cov-report=term-missing --cov-fail-under=85"
testpaths = ["tests"]
markers = ["integration: end-to-end tests with real rendering"]

[tool.mypy]
ignore_missing_imports = true
no_strict_optional = true
warn_unused_configs = true
warn_redundant_casts = true
